requirements.txt
discord.py==2.2.3
pytest==7.3.1
requests==2.31.0
pylint==2.17.4